* Fixes NB#57580 	* Fixed some issues in a text conversion function 	* Do not dimm Email->Details if the folder view has the focus
pmo-trunk-r2020

+++ src/modest-text-utils.c
@@ -1121,11 +1121,11 @@ modest_text_utils_get_display_size (guint64 size)
 	else if (KB <= size && size < 100 * KB)
 		return g_strdup_printf (_FM("sfil_li_size_1kb_99kb"), size / KB);
 	else if (100*KB <= size && size < MB)
-		return g_strdup_printf (_FM("sfil_li_size_100kb_1mb"), size / MB);
+		return g_strdup_printf (_FM("sfil_li_size_100kb_1mb"), (float) size / MB);
 	else if (MB <= size && size < 10*MB)
-		return g_strdup_printf (_FM("sfil_li_size_1mb_10mb"), size / MB);
+		return g_strdup_printf (_FM("sfil_li_size_1mb_10mb"), (float) size / MB);
 	else if (10*MB <= size && size < GB)
 		return g_strdup_printf (_FM("sfil_li_size_10mb_1gb"), size / MB);
 	else
-		return g_strdup_printf (_FM("sfil_li_size_1gb_or_greater"), size / GB);	
+		return g_strdup_printf (_FM("sfil_li_size_1gb_or_greater"), (float) size / GB);	
 }
